PSTRNG-28 : Add Environment Loading function to bashrc, and set gitattributes to get around windows based repositories.

+++ deploy/load_env.sh
@@ -2,20 +2,20 @@
  
 # CP ENVIRONMENT
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-arr=(${SCRIPT_DIR}/std_unix_env/*);
 
-# i=0
-# while read line
-# do
-#     arr[ $i ]="$line"        
-#     (( i++ ))
-# done < <(ls -ls ~/sync/std-unix-env/)
+
+if [[  -d ~/workspace/deploy && ! -f ~/.alias ]]; then
+  # CP ENVIRONMENT
+  arr=(${SCRIPT_DIR}/std_unix_env/*);
+  
+  for f in "${arr[@]}"; do
+    #cp "$f" "~/.$f"
+    echo -e "Copying $f."
+    echo -e "Filename : ${f##*/}"
+    filename=${f##*/}
+    cp -R $f ~/.$filename
+  done
+fi
 
 # iterate through array using a counter
-for f in "${arr[@]}"; do
-  #cp "$f" "~/.$f"
-  echo -e "Copying $f."
-  echo -e "Filename : ${f##*/}"
-  filename=${f##*/}
-  cp -R $f ~/.$filename
-done
+
